Move persist_job_config method from base service.
Method is only used in pipeline, obs, and credentials services.
