+++ mash/services/mash_service.py
@@ -33,7 +33,6 @@ from mash.mash_exceptions import (
     MashRabbitConnectionException,
     MashLogSetupException
 )
-from mash.utils.json_format import JsonFormat
 
 
 class MashService(object):
@@ -201,19 +200,6 @@ class MashService(object):
         self.channel.basic.consume(
             callback=callback, queue=queue
         )
-
-    def persist_job_config(self, config):
-        """
-        Persist the job config file to disk for recoverability.
-        """
-        config['job_file'] = '{0}job-{1}.json'.format(
-            self.job_directory, config['id']
-        )
-
-        with open(config['job_file'], 'w') as config_file:
-            config_file.write(JsonFormat.json_message(config))
-
-        return config['job_file']
 
     def restart_jobs(self, callback):
         """
