+++ mash/services/obs/service.py
@@ -259,3 +259,16 @@ class OBSImageBuildResultService(MashService):
             'ok': True,
             'message': 'Job started'
         }
+
+    def persist_job_config(self, config):
+        """
+        Persist the job config file to disk for recoverability.
+        """
+        config['job_file'] = '{0}job-{1}.json'.format(
+            self.job_directory, config['id']
+        )
+
+        with open(config['job_file'], 'w') as config_file:
+            config_file.write(JsonFormat.json_message(config))
+
+        return config['job_file']
