+++ mash/services/pipeline_service.py
@@ -602,6 +602,19 @@ class PipelineService(MashService):
         except Exception:
             pass
 
+    def persist_job_config(self, config):
+        """
+        Persist the job config file to disk for recoverability.
+        """
+        config['job_file'] = '{0}job-{1}.json'.format(
+            self.job_directory, config['id']
+        )
+
+        with open(config['job_file'], 'w') as config_file:
+            config_file.write(JsonFormat.json_message(config))
+
+        return config['job_file']
+
     def start(self):
         """
         Start pipeline service.
